requirements.txt
boto==2.45.0
pandas==0.19.2
plotly==2.0.6
pytz==2016.10
radiotherm==1.3
